# Improve performance of BigInteger.GCD