--- a/src/System.Runtime.Numerics/src/System/Numerics/BigIntegerCalculator.BitsBuffer.cs
+++ b/src/System.Runtime.Numerics/src/System/Numerics/BigIntegerCalculator.BitsBuffer.cs
@@ -119,6 +119,57 @@ namespace System.Numerics
                 }
             }
 
+            [SecuritySafeCritical]
+            public unsafe void Reduce(ref BitsBuffer modulus)
+            {
+                // Executes a modulo operation using the divide operation.
+                // Thus, no need of any switching here, happens in-line.
+
+                if (_length >= modulus._length)
+                {
+                    fixed (uint* b = _bits, m = modulus._bits)
+                    {
+                        Divide(b, _length,
+                               m, modulus._length,
+                               null, 0);
+                    }
+
+                    _length = ActualLength(_bits, modulus._length);
+                }
+            }
+
+            public void Overwrite(ulong value)
+            {
+                Debug.Assert(_bits.Length >= 2);
+
+                if (_length > 2)
+                {
+                    // ensure leading zeros
+                    Array.Clear(_bits, 2, _length - 2);
+                }
+
+                uint lo = (uint)value;
+                uint hi = (uint)(value >> 32);
+
+                _bits[0] = lo;
+                _bits[1] = hi;
+                _length = hi != 0 ? 2 : lo != 0 ? 1 : 0;
+            }
+
+            public void Overwrite(uint value)
+            {
+                Debug.Assert(_bits.Length >= 1);
+
+                if (_length > 1)
+                {
+                    // ensure leading zeros
+                    Array.Clear(_bits, 1, _length - 1);
+                }
+
+                _bits[0] = value;
+                _length = value != 0 ? 1 : 0;
+            }
+
             public uint[] GetBits()
             {
                 return _bits;
@@ -127,6 +178,24 @@ namespace System.Numerics
             public int GetSize()
             {
                 return _bits.Length;
+            }
+
+            public int GetLength()
+            {
+                return _length;
+            }
+
+            public void Refresh(int maxLength)
+            {
+                Debug.Assert(_bits.Length >= maxLength);
+
+                if (_length > maxLength)
+                {
+                    // ensure leading zeros
+                    Array.Clear(_bits, maxLength, _length - maxLength);
+                }
+
+                _length = ActualLength(_bits, maxLength);
             }
 
             private void Apply(ref BitsBuffer temp, int maxLength)