… src/System.Runtime.Numerics/src/System/Numerics/BigIntegerBuilder.cs
… src/System.Runtime.Numerics/tests/BigInteger/gcd.cs