- In order to consolidate the big integer calculation code, Lehmer's algorithm gets some refactoring and a new home within the recently added BigIntegerCalculator.
- The remaining code of BigIntegerBuilder to convert a big integer value to a double value moves to BigInteger, where the other conversion code lives too.
- BigIntegerBuilder is finally dead now, it gets dropped.
- NumericsHelpers contains dead code now, it gets some cleanup.

The performance of *GCD* improves a bit after these changes:

**GCD**

| # of bits | # of vals | before ms | after ms |
|----------:|----------:|----------:|---------:|
|        16 | 1,000,000 |       126 |       81 |
|        64 | 1,000,000 |       369 |      328 |
|       256 |   100,000 |       238 |      236 |
|     1,024 |   100,000 |      1113 |     1077 |
|     4,096 |    10,000 |       675 |      581 |
|    16,384 |     1,000 |       629 |      465 |
|    65,536 |       100 |       816 |      557 |

#1307