@stephentoub 2015-09-18
Thanks, @axelheer.  Sorry for the delay.  I'll review soon.