@KrzysztofCwalina 2015-09-18
This looks good to me. Thanks again for doing great work to improve this component.