@stephentoub 2015-09-20
I left a few comments on the change about places it looks like we're missing test coverage.  Other than that, LGTM.