@axelheer 2015-09-20
@stephentoub thanks, I'll check those return statements. The break statements are the same as the current implementation, I'm not sure how to find values hitting these edge cases, but I'll think about it.