@axelheer 2015-09-20
@stephentoub I found some values to cover all those statements and finally deleted two branches, which don't pay off anyway.